DRY'd out the datasources
