+++ lib/components/Datasource.react.js
@@ -2,9 +2,9 @@
 
 var React = require('react');
 var Reflux = require('reflux');
-var _ = require('underscore');
+
 var UtilityMixin = require('../mixins/UtilityMixin');
-var DSManager = require('../services/DataStoreManager');
+var DataMixin = require('../mixins/DataMixin');
 
 /**
  * Datasource
@@ -51,34 +51,12 @@ var Datasource = React.createClass({
 
   mixins: [
     Reflux.ListenerMixin,
-    UtilityMixin
+    UtilityMixin,
+    DataMixin
   ],
-  
-  _buildProps() {
-    var props = _.extend({}, this.props);
-    delete props.children;
-    delete props.config;
 
-    var data = this.state.data;
-    var actions = this.dataSource && this.dataSource.listenables;
-    return _.extend(props, { data, actions });
-  },
-
-  getInitialState() {
-    return {
-      data: this.props.data
-    };
-  },
-
-  _prepareDatasource(props) {
-    var { config } = props;
-    this.dataSource = DSManager.instance().getStore(config);
-    this.listenTo(this.dataSource, this.onStoreChange);
-  },
-
-  _reloadCollection(id) {
-    var data = this.dataSource.get(id);
-
+  _reloadCollection(props) {
+    var data = this.dataSource.get(props.objectId);
     this.setState({ data });
   },
 
@@ -87,26 +65,13 @@ var Datasource = React.createClass({
       this.stopListeningToAll();
       this._prepareDatasource(nextProps);
     } else if (nextProps.objectId !== this.props.objectId) {
-      this._reloadCollection(nextProps.objectId);
+      this._reloadCollection(nextProps);
     }
   },
 
   componentDidMount() {
     this._prepareDatasource(this.props);
-    this._reloadCollection(this.props.objectId);
-  },
-
-  onStoreChange(type, payload) {
-    this._reloadCollection(this.props.objectId);
-  },
-
-  render() {
-    var children = this._cloneWithProps(this._buildProps());
-    return(
-      <div>
-        {children}
-      </div>
-    );
+    this._reloadCollection(this.props);
   }
 });
