+++ lib/components/DependentDatasource.react.js
@@ -2,9 +2,9 @@
 
 var React = require('react');
 var Reflux = require('reflux');
-var _ = require('underscore');
+
 var UtilityMixin = require('../mixins/UtilityMixin');
-var DSManager = require('../services/DataStoreManager');
+var DataMixin = require('../mixins/DataMixin');
 
 /**
  * DependentDatasource
@@ -60,31 +60,9 @@ var DependentDatasource = React.createClass({
 
   mixins: [
     Reflux.ListenerMixin,
-    UtilityMixin
+    UtilityMixin,
+    DataMixin
   ],
-  
-  _buildProps() {
-    var props = _.extend({}, this.props);
-    delete props.children;
-    delete props.config;
-
-    var data = this.state.collection;
-    var actions = this.dataSource && this.dataSource.listenables;
-    return _.extend(props, { data, actions });
-  },
-
-  getInitialState() {
-    return {
-      collection: this.props.collection
-    };
-  },
-
-  _prepareDatasource(props) {
-    var { config } = props;
-
-    this.dataSource = DSManager.instance().getStore(config);
-    this.listenTo(this.dataSource, this.onStoreChange);
-  },
 
   _updateDynamicView(props) {
     var data = props.data;
@@ -98,8 +76,8 @@ var DependentDatasource = React.createClass({
   },
 
   _reloadCollection(props) {
-    var collection = this.dataSource.getCollection(props.tag);
-    this.setState({ collection });
+    var data = this.dataSource.getCollection(props.tag);
+    this.setState({ data });
   },
 
   componentWillReceiveProps(nextProps) {
@@ -125,19 +103,6 @@ var DependentDatasource = React.createClass({
   componentDidMount() {
     this._prepareDatasource(this.props);
     this._updateDynamicView(this.props);
-  },
-
-  onStoreChange(type, payload) {
-    this._reloadCollection(this.props);
-  },
-
-  render() {
-    var children = this._cloneWithProps(this._buildProps());
-    return(
-      <div>
-        {children}
-      </div>
-    );
   }
 });
