+++ lib/components/FilteredDatasource.react.js
@@ -2,9 +2,9 @@
 
 var React = require('react');
 var Reflux = require('reflux');
-var _ = require('underscore');
+
 var UtilityMixin = require('../mixins/UtilityMixin');
-var DSManager = require('../services/DataStoreManager');
+var DataMixin = require('../mixins/DataMixin');
 
 /**
  * FilteredDatasource
@@ -56,39 +56,19 @@ var FilteredDatasource = React.createClass({
 
   getDefaultProps: function() {
     return {
-      idAttr: "id",
-      storeType: "unique",
+      config: {
+        idAttr: "id",
+        storeType: "unique"
+      },
       data: []
     };
   },
 
   mixins: [
     Reflux.ListenerMixin,
-    UtilityMixin
+    UtilityMixin,
+    DataMixin
   ],
-  
-  _buildProps() {
-    var props = _.extend({}, this.props);
-    delete props.children;
-    delete props.config;
-
-    var data = this.state.collection;
-    var actions = this.dataSource && this.dataSource.listenables;
-    return _.extend(props, { data, actions });
-  },
-
-  getInitialState() {
-    return {
-      collection: this.props.collection
-    };
-  },
-
-  _prepareDatasource(props) {
-    var { config } = props;
-
-    this.dataSource = DSManager.instance().getStore( config );
-    this.listenTo(this.dataSource, this.onStoreChange);
-  },
 
   _registerView(props) {
     var { tag } = props;
@@ -97,8 +77,8 @@ var FilteredDatasource = React.createClass({
   },
 
   _reloadCollection(props) {
-    var collection = this.dataSource.getCollection(props.tag);
-    this.setState({ collection });
+    var data = this.dataSource.getCollection(props.tag);
+    this.setState({ data });
   },
 
   componentWillReceiveProps(nextProps) {
@@ -117,19 +97,6 @@ var FilteredDatasource = React.createClass({
   componentDidMount() {
     this._prepareDatasource(this.props);
     this._registerView(this.props);
-  },
-
-  onStoreChange(type, payload) {
-    this._reloadCollection(this.props);
-  },
-
-  render() {
-    var children = this._cloneWithProps(this._buildProps());
-    return(
-      <div>
-        {children}
-      </div>
-    );
   }
 });
